Move dashboards to each beat
All dashboards from beats-dashboards were moved to their specific beats.
Two new functions were introduced in the Makefile to manage import / export of the files from Kibana:

* `make kibana-export`
* `make kibana-import`

If `KIBANA_ES` is not running on localhost, KIBANA_ES has to be overwritten. Be aware that the current implementation is to simplify development of dashboards an is not intended to serve yet for the deployment of dashboards of the users. This should be implemented in a second step.

General and reusable dashboard components were moved to libbeat. These should be in a second step automatically move to each beat when running `make update` similar to what is done with `libbeat.yml`.

Further:
* Basic Redis Dashboard in modules were added
* Kibana dir in modules is called .kibana to prevent naming conflicts
* README for scripts folder extended
* Add `make beats-dashboards` to generate dashboards for https://github.com/elastic/beats-dashboards repo
* Add `make kibana` to metricbeat to collect kibana files from modules

Currently I'm not happy with the naming of etc/kibana but we should move forward with this and find a better solution on the way.

This change is also needed for Metricbeat where the dashboards will be part of each module. This can potentially also be used for packetbeat modules.