Dynamic fields path in fields generator script
For nested fields, it was necessary to either define the name with the dot nation based on its hierarchy or using the path variable. The generation script was now adjusted to automatically keep track of the hierarchy and generate the names correctly. This simplifies the fields.yml.

In addition it does not require knowledge of the full hierarchy in nested documents. This is especially useful for metricset fields docs as they should not require to have information about the path of their module.

* Packetbeat and Topbeat files were adjusted to meet the new logic.
* A prefix flag for docs was introduced to manually set the [float] flag which is needed in topbeat because of the nesting.
* Fix typo in filebeat config. Closes #987
* Adjust system test scripts to support new format