Refactor the API for implementing Modules and MetricSets (#1549)
- Refactor the API for implementing Modules and MetricSets.
- Integration test cleanup to make all of them consistent with each
other.
- Refactor goroutines for workers to use one per host. Previously it
created new a goroutine on every fetch. There was a cap on the maximum
number of concurrent goroutines allowed per MetricSet. If there were
more hosts than the cap some fetches would not execute. Or if a host
became unresponsive then after a number of successive fetches, the
failed host could consume all the available goroutines for a MetricSet.
With this change an unresponse host will be isolated.
- Fix all golint warnings.
- Adjust period used in system tests to account for file output flush
slowness when running inside Docker.
- Separate github.com/elastic/beats imports from all other imports in Metricbeat.
- Add package github.com/pkg/errors from version cd6e0b425dfeaa184e93ed993aaee9dea03dc14d.