Add script to generate MetricSet (#1738)
Adding a MetricSet requires a create several files with a predefined file structure for automation. To simplify the task of creating a MetricSet and make sure the correct structure is used, this task is now automated with a script. The script can be called with `make create-metricset MODULE=modulename METRICSET=metricsetname`. This generates all files including template, config and docs. After `make imports` must be called to update the import list and `make collect` to aggregate all the new data. The MetricSet can directly be compiled and run with the default content.

This script can also be used in the metricbeat-generator, means someone creating multiple metricsets has also the tooling available. In addition, this simplifies the metricbeat-generator as most of the templates which change frequently are in the beats repository.

As the templates used for the generation are the same as shown in the developer guide, it can also be used in the docs. This makes also the docs always up-to-date. One example was added to the docs. The go files have `.tmpl` appended to make sure they are not compiled.

This fixes also a fleaky filebeat test which was caused by inode reuse.