Unify version handling
The new file `libbeat/docs/version.asciidoc` contains a single line with the
`stack-version` variable. All other version variables reference this one. This
makes it easy to automate version changing.

`make update` is needed to copy the file to the individual Beats. It would have
been possible to include directly the file from libbeat, but that would have had
a big disadvantage: the docs only get rebuild on file changes in the "book", so
changing the version in libbeat wouldn't trigger a docs rebuild on the public
website.

Also, it's kind of nice to have to run `make update` shortly before each release,
that makes it sure we won't forget.