Fix Filebeat rotation issue (#1954)
In case a file was renamed after the state was read but before the file was opened by the harvester it could happen that the wrong file was opened. This lead to the issue, that the wrong file was read including reporting the wrong state for a file. In general this was rather unlikely to happen, but can happen when scan_frequency is very low and number of files rotated is very high.

This issue is fixed with the harvester now compares again the file after opening with the state given from the prospector. In case the state is not identical with the file, the harvester aborts. This means that the file will only be picked up again after scan_frequency but potential errors are prevented.