Add state for ignore_older files (#2859)
Previously if a file was falling under ignore_older on startup and no previous state was found, also no state was persisted. This was important in 1.x releases to prevent the registry file to contain too many files as no cleanup was possible. With the new clean_* options the registry file can now be cleaned up. For consistency all files that fall under ignore_older should have a state, even if they are not crawled before.

This change is also required to move tail_files under the prospector (https://github.com/elastic/beats/issues/2613). Otherwise files which fall under `ignore_older` the state could not be set to the end of the file.