Filebeat module system/integration tests (#3214)
* Added system/integration tests for the Filebeat modules

The tests are taking all the logs from the modules `test` folders,
run filebeat on them (loading the template and the pipeline), then
do a search against Elasticsearch. On the result from Elasticsearch,
the test checks that all fields are documented.

In addition, the module writer can proved `-expected` files for each
log file, in which case we test the contents as well.

Part of #3159.

* Run system tests in docker so that the test env is available

Also adds the ingest plugins to our snapshot file, which are needed
for the modules tests.

* Skip permission test when run as root

* Don't run integration tests on OSX

Also install elasticsearch-py for appveyor.