Make -setup load the Beat dashboards (#3506)
With this change, `filebeat -setup` loads the sample Kibana dashboards,
being the rough equivalent of running: `scripts/import_dashboards && filebeat`.
This works for all Beats, not only Filebeat. The `import_dashboards` program
is kept for now, to avoid braking compatibility. Code wise, the two share
most of the code, which was separated in a sub-package.

All flags available for the `import_dashboard` are available as go-ucfg options.
For example, to load the dashboards from a directory you can do:

    filebeat -e -setup -E "dashboards.dir=_meta/kibana"

Or do load the snapshot version of the dashboards:

    filebeat -e -setup -E "dashboards.snapshot=true"

This also changes the pipeline loading to happen automatically regardless of
whether `-setup` was used or not.

Part of #3159.

Contains updates for the docs and changelog.