Fix packaging which broke because of asciidocs comments (#3825)
The version.asciidoc is used during packaging to read out the version number. Recently a comment was introduce in the file which broke the packaging. In before_build.sh the content of the file is copied over and converted to yaml, see https://github.com/elastic/beats/blob/master/dev-tools/packer/xgo-scripts/before_build.sh#L58. As these lines to not seem to be required, I removed them. Alternatively the sed script could be adapted which would be more complex.