Clarify docs around setting the index and @metadata fields (#3866)
This resolves issue #3558 by clarifying how the contents of the @metadata field are used for setting the index. I've also done a little cleanup in general of the Logstash config docs.

As a little side project, I removed "Configuration" from the end of the topic titles for the reference content because the navigation is a lot easier to read (IMO) without "Configuration" repeated.