Add audit module with kernel metricset to Metricbeat (#4178)
The `kernel` metricset receives audit events from the Linux Audit Framework that
is a part of the Linux kernel.

This metricset establishes a subscription to the `kernel` to receive the events
as they occur.

The Linux Audit Framework can send multiple messages for a single auditable
event. This metricset will combine all of the data from each of the messages
into a single event.

Messages for one event can be interleaved with messages from another event.
This metricset will buffer the messages in order to combine related messages
into a single event even if they arrive interleaved or out of order.

It can resolve numeric UIDs and GIDs in the event to their names.

Syscalls, architectures, and socket addresses are translated from numeric
representations to a human readable form.

Based on the audit message type, a `category` and descriptive `action` are
added to the event.