Start cleaning up condition tests (#5960)
This simplifies part of the condition tests to use table tests. Like this the expected result and the condition itself are close together in the code which makes it easier to read.

Cleanup was triggered when looking at https://github.com/elastic/beats/pull/5954 and realised it hard to follow in a diff.