Load ML using Kibana API (#6604)
Since Kibana 6.1 new ML API is available to load jobs, datafeeds, etc. Thus, it's not necessary to store ML configs in Beats and load them using setup.

From now on if Filebeat connects to a Kibana 6.1 or newer the Kibana API is called. If Kibana is older than 6.1, Filebeat loads the ML config from the repo, as it's done in previous versions.