Rename metrics/logs builders to `hints` (#6621)
This change renames current `logs` and `metrics` builders to `hints`. These builders are beat specific and are minded to create configs based on hints (from labels or annotations). Config would then look like this:

```
 metricbeat.autodiscover:
   providers:
     - type: kubernetes
       builders:
        - type: hints
```

It also introduces a shortcut to configure just the `hints` builder, which should cover most use cases for common users:

```
metricbeat.autodiscover:
  providers:
    - type: kubernetes
      hints.enabled: true
```

No changelog is needed as this is updating hints based autodiscover, which is unreleased