Add seccomp whitelist policy
This PR adds to the seccomp enhancement in #6832 by adding a whitelist policy.

The policies are derived from a combination of static analysis of the object code and runtime profiling (to see what the cgo linked libraries are doing). The policies also account for the system calls used while running the system tests with coverage and race detection.

Closes #5213