Cherry-pick #7476 to 6.x: Overwrite feature (#8575)
Cherry-pick of PR #7476 to 6.x branch. Original message: 

Allow to Overwrite an existing feature in the registry and add package
level function to deal bundle that need to overwrite feature.

This is require by the feature defined in the `add_kubernetes_metadata`
processors, the core ships with a set of default options for the feature
but a specific beats can decide to provide another set of default options.