Enable set syslog host for Filebeat HAProxy module (#9366) (#9403)
The syslog input for the Filebeat HAProxy module binds by default to
localhost:9001. It is possible to configure the port through the
variable var.syslog_port, but it is not possible for the
syslog_host. This can be useful when running Filebeat inside a Docker
container. Since it is not easy to forward a port from the host to
localhost inside the Docker container, it restricts the ability to use
Filebeat in a Docker container to log HAProxy running on a remote
instance. With this commit, we make it possible to configure the host
that the syslog input for the HAProxy binds to through a variable
var.syslog_host. This makes it possible to have HAProxy running on a
remote instance (e.g., a FreeBSD 11.2 host), and configure that instance
to remote log to a Filebeat instance running inside a Docker container.

(cherry picked from commit bc7640f4b25afc192dafd80ddf130ef269737c84)

Co-authored-by: Jason Tedor <jason@tedor.me>