Backport: Generate fields.go for netflow (#9740)
* Backport fields generation code

Cherry-pick select parts of eb1739e30c087fb13ecc5af0c699d6aae0f2bc59

* Generate fields.go for netflow input

* Generate fields.go for netflow input

Update x-pack/filebeat to generate a fields.go for each input type. And update the code that
generates the include/list.go to import each input.

* Fix netflow exporter version datatype

It was "int" which is not a valid type. Changed it to long rather than integer because we use that everywhere for numbers.

Cherry-pick parts of 01ae53802c8da17724326da07d142a94ab716810 (#9686)

* Update x-pack/filebeat to use the xpack.mk

The existing Filebeat Makefile could not be reused because generate_imports.py is not portable.

* Add clean and exportDashboard targets

* createDir in GenerateModuleReferenceConfig

* Remove makefile hack to test x-pack/filebeat