Cherry-pick to 6.x: Ignore timestamp in redis, haproxy and system filebeat module (#9855) (#9857)
* Ignore timestamp in redis, haproxy and system filebeat module (#9855)

* Ignore timestamp in redis, haproxy and system filebeat module

* Fixing formatting

(cherry picked from commit b9a20f7ba0b42c98c223a9d34216957a1922372e)

* Change to use event.dataset instead

* Update golden files for icinga/startup test to exclude timestamp (#9506)

The icing/startup log file does not contain a timestamp. Because of this the timestamp from filebeat is taken. During the generation of golden files still a timestamp was added but every skipped on comparison. Instead now the timestamp is not added to the generated file anymore to now show a diff each time GENERATE is run.

* Regenerate expectations for logs without timestamps (#9862)

* Rerun icinga redis haproxy and system modules with GENERATE=1