Cherry-pick #10995 to 6.7: Include ip and boolean type when generating index pattern (#11091)
Cherry-pick of PR #10995 to 6.7 branch. Original message: 

A number of fields in our `fields.yml` files now are `type: ip` (e.g. `source.ip`) and `type: boolean`, but the code generating index patterns does not know about these types yet and so does not add a `type` to the index pattern for those fields at all. This leads to errors in Kibana when looking at dashboards that contain references to those fields.