Detect if the configured and the migrated registry paths are the same (#10982) (#11085)
Cherry-pick of PR #10982 to 6.7 branch. Original message: 

The E2E test of starting Journalbeat with an existing registry broke when I have added support and tests for `include_matches`. Thus, when I added the migration of the registry file, the issue was not detected.
Now both the code and the test are fixed.